Fixed and hashed passwords
Made the changes to the database so that the passwords are hashes in SHA2. NOW must make the notes to have new passwords hashed in SHA2 too so that when new passwords are made they're saved NOT as strings.

Now with this type of hashing, login passwords can be made into SHA2 and then compared.