Made changes in DB to allow for Salt (zest) and making some alterations to the signup so it will actully sen correct INSERT mySQL commands.
Database is setup with no Normal FOrms and I believe that's causing it to deny the Insert data from the FrontEnd.

Pascual Sahar and I all worked from 1pm to about 4pm trying to slave this, but...  We DID get the connect.php to be a header in every page and have the queries be correct, such that the SELECT * FROM ...  works as intended and the hashing + salt functions work nicely.   +Database now has the varchar(length) sutiable for the elongated characters being pushed into it.

But to reiterate, believe Accouts and Users table need to be Normalized befor data can be pushed into DB.  -Pascual + Sahar also found issues in the HTML where the buttons are having issues so that may also be a/the issue.  Still working on that...