Made changes to backend.
Hopefully these changes will make it possible for pushing data to backend from front end easier. Believe there were anomalies of the "update" variety not allowing the front-end to function properly.

The .JPG images are the work that gfuentess and I  did to see if we can revamp the DB.

Should hopefully remedy the update and delete anomalies

Co-Authored-By: gfuentesherr <60154823+gfuentesherr@users.noreply.github.com>